Introduce a persistent sidebar shared across all tabs
Pulls the logo, news feed, install/version controls, main action
button, and social links out of the Home/Community tab content and
into a new SidebarView that lives beside the TabControl instead of
inside it, so it stays visible and identical no matter which tab is
selected. News ownership moves up to MainWindowViewModel since it's
no longer tab-specific. Also drops the now-redundant logo from
TabControl's header template since the sidebar carries it instead.
