+++ Ksp2Redux.Tools.Launcher/ViewModels/Community/CommunityTabViewModel.cs
@@ -4,10 +4,8 @@ using Ksp2Redux.Tools.Launcher.ViewModels.Shared;
 
 namespace Ksp2Redux.Tools.Launcher.ViewModels.Community;
 
-public partial class CommunityTabViewModel(INewsItemCollectionService newsCollectionService, INewsService newsService) : ViewModelBase
+public partial class CommunityTabViewModel(INewsService newsService) : ViewModelBase
 {
-    public NewsCollectionViewModel NewsCollectionViewModel { get; set; } = new(newsCollectionService.NewsCollection);
-
     private int SelectedNewsId
     {
         get;
