+++ Ksp2Redux.Tools.Launcher/Views/Shared/NewsCollectionView.axaml
@@ -11,11 +11,12 @@
                    FontSize="24" TextAlignment="Center" Margin="0,16,0,8">
             LATEST NEWS
         </TextBlock>
-        <Border Grid.Row="1" Grid.Column="0" Background="#802D2D2D" CornerRadius="12"
+        <Border Grid.Row="1" Grid.Column="0" Background="{StaticResource SurfacePanelTranslucentBrush}" CornerRadius="12"
                 ClipToBounds="True"
                 Margin="0,0,0,0"
                 VerticalAlignment="Stretch">
-            <ScrollViewer Width="260" Height="340"
+            <ScrollViewer Width="260"
+                          VerticalAlignment="Stretch"
                           ClipToBounds="True"
                           VerticalScrollBarVisibility="Auto">
                 <ItemsControl Margin="20,16, 20,16" ItemsSource="{Binding NewsCollection}"
